Code refactoring, add Comments
